almost-working 32-bit mixing

+++ include/gba/m4a_internal.h
@@ -214,7 +214,7 @@ struct SoundInfo
     ExtVolPitFunc ExtVolPit;
     u8 gap2[16];
     struct SoundChannel chans[MAX_DIRECTSOUND_CHANNELS];
-    s8 pcmBuffer[PCM_DMA_BUF_SIZE * 2];
+    s32 pcmBuffer[PCM_DMA_BUF_SIZE * 2];
 };
 
 struct SongHeader
